refactor: apps/packages layout; add A+ template sync; update aliases, vite, tailwind, docs
