+++ apps/web/public/config/cardPacks.json
