+++ apps/web/public/config/cards/base/base.json
