+++ apps/web/public/illustrations/default.webp
