+++ apps/web/public/illustrations/player_male_ecologist.webp
