+++ apps/web/src/games/demo-with-backend/components/DebugInfo.tsx
