+++ apps/web/src/games/demo-with-backend/hooks/useAuth.ts
