--- src/carrot/config.ts
@@ -1,3 +1,0 @@
-export const config = {
-  // 未来可以在这里添加模板引擎的通用配置
-};
