+++ sync.manifest
@@ -1,0 +1,13 @@
+configs/
+packages/ui/
+packages/services/
+scripts/
+apps/web/tailwind.config.js
+.editorconfig
+.eslintrc.cjs
+postcss.config.js
+tailwind.config.js
+tsconfig.json
+vite.config.ts
+README.md
+docs/
