+++ vite.config.ts
@@ -16,15 +16,22 @@ export default defineConfig(({ command, mode }) => {
     GITHUB_PAGES: process.env.GITHUB_PAGES
   })
 
+  const appRoot = path.resolve(__dirname, 'apps/web');
+
   return {
+    root: appRoot,
     plugins: [react()],
     base: process.env.GITHUB_PAGES ? `/${pkg.name}/` : './',
     resolve: {
       alias: {
-        '@': path.resolve(__dirname, './src'),
+        '@': path.resolve(__dirname, './apps/web/src'),
+        '@ui': path.resolve(__dirname, './packages/ui/src'),
+        '@services': path.resolve(__dirname, './packages/services/src'),
       },
     },
     build: {
+      outDir: path.resolve(__dirname, 'dist'),
+      emptyOutDir: true,
       assetsDir: 'assets',
       rollupOptions: {
         output: {
@@ -36,7 +43,6 @@ export default defineConfig(({ command, mode }) => {
       'import.meta.env.VITE_INCLUDE_START_PACK': JSON.stringify(env.VITE_INCLUDE_START_PACK),
       'import.meta.env.VITE_BACKEND_URL': JSON.stringify(env.VITE_BACKEND_URL),
       'import.meta.env.VITE_PUBLIC_URL': JSON.stringify(env.VITE_PUBLIC_URL),
-      'import.meta.env.BASE_URL': JSON.stringify(process.env.GITHUB_PAGES ? `/${pkg.name}/` : './')
     }
   }
 })
